Move tracks & friends into viz/
