+++ src/main/VisualizationWrapper.js
@@ -9,7 +9,7 @@ import type {VizWithOptions} from './types';
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import d3utils from './d3utils';
+import d3utils from './viz/d3utils';
 import _ from 'underscore';
 import d3 from '../lib/minid3';
