+++ src/main/viz/canvas-utils.js
