+++ src/test/TiledCanvas-test.js
@@ -6,7 +6,7 @@ import {expect} from 'chai';
 import _ from 'underscore';
 
 import Interval from '../main/Interval';
-import TiledCanvas from '../main/TiledCanvas';
+import TiledCanvas from '../main/viz/TiledCanvas';
 
 describe('TiledCanvas', function() {
   describe('getNewTileRanges', function() {
